muestra la cantidad de productos del carrito
