+++ src/Components/Header.js
@@ -9,9 +9,10 @@ import { Link as LinkRouter, useNavigate } from "react-router-dom";
 import { useUserLogoutMutation } from "../Features/usersAPI";
 import { reload } from "../Features/reloadSlice";
 import { deleteCredentials } from "../Features/usersSlice";
-import Button from 'react-bootstrap/Button';
-import Offcanvas from 'react-bootstrap/Offcanvas';
-import BasketCart from "./BasketCart";
+// import Button from 'react-bootstrap/Button';
+// import Offcanvas from 'react-bootstrap/Offcanvas';
+// import BasketCart from "./BasketCart";
+import BtnCart from "./BtnCart";
 
 export default function Header({ name, ...props }) {
   const userData = useSelector((state) => state.auth.user);
@@ -21,9 +22,9 @@ export default function Header({ name, ...props }) {
   const logged = useSelector((state) => state.auth.logged);
   const [singOutUser] = useUserLogoutMutation()
 
-  const [show, setShow] = useState(false);
-  const handleClose = () => setShow(false);
-  const handleShow = () => setShow(true);
+  // const [show, setShow] = useState(false);
+  // const handleClose = () => setShow(false);
+  // const handleShow = () => setShow(true);
 
   const handleOpenMenu = () => {
     if (open === true) {
@@ -71,18 +72,7 @@ export default function Header({ name, ...props }) {
               <img src={logged ? userData.photo : 'user.png'} onClick={handleOpenMenu} />
             </div>
             <div className="headerCart">
-              <button variant="primary" onClick={handleShow} className="me-2">
-                <img src='https://cdn-icons-png.flaticon.com/512/8070/8070478.png' alt='icon' />
-              </button>
-              <Offcanvas placement='end' show={show} onHide={handleClose} {...props}>
-                <Offcanvas.Header closeButton>
-                  <Offcanvas.Title>CARRITO DE COMPRAS</Offcanvas.Title>
-                  <hr/>
-                </Offcanvas.Header>
-                <Offcanvas.Body>
-                  <BasketCart/>
-                </Offcanvas.Body>
-              </Offcanvas>
+              <BtnCart/>
             </div>
           </div>
           {logged ? (
